+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Downloads Proxy version 2 
+  [#242](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/242)
 * Note about data currency in WMS usage instructions
   [#233](https://gitlab.data.bas.ac.uk/MAGIC/add-metadata-toolbox/-/issues/233)